// frontend/src/openapi/generated/model/categories-and-memo-count.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cats example
 * The cats API description
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Memo } from './memo';
// May contain unused imports in some cases
// @ts-ignore
import { Tag } from './tag';

/**
 * 
 * @export
 * @interface CategoriesAndMemoCount
 */
export interface CategoriesAndMemoCount {
    /**
     * 
     * @type {number}
     * @memberof CategoriesAndMemoCount
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof CategoriesAndMemoCount
     */
    'cateName': string | null;
    /**
     * 
     * @type {number}
     * @memberof CategoriesAndMemoCount
     */
    'userId': number;
    /**
     * 
     * @type {Array<Memo>}
     * @memberof CategoriesAndMemoCount
     */
    'memo': Array<Memo>;
    /**
     * 
     * @type {Array<Tag>}
     * @memberof CategoriesAndMemoCount
     */
    'tag': Array<Tag>;
    /**
     * 
     * @type {number}
     * @memberof CategoriesAndMemoCount
     */
    'memoCount'?: number;
}