// frontend/src/openapi/generated/model/cate-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cats example
 * The cats API description
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CateInput
 */
export interface CateInput {
    /**
     * 
     * @type {number}
     * @memberof CateInput
     */
    'cateId': number;
    /**
     * 
     * @type {string}
     * @memberof CateInput
     */
    'cateName': string;
}